Remove the need for a relationship between servers and new_servers.
Previously, servers and new_servers had to share a prefix, because
MemcachedStoreView returned indexes into an array of memcached
connections (so if your servers[0] didn't match new_servers[0], read
attempts would go to a different memcached and be lost).

This change makes MemcachedStoreView return the address of the memcached
server directly, rather than an index, allowing for new_servers to have
any (or no) overlap with servers.