Cassandra Connection pool changes
If the cassandra process goes down and comes back up, connections to that
process are no longer valid and should be torn down and recreated.

So this commit adds a member variable to the connection pool. If true, when a
connection is released but marked not to be returned to the pool, then all
existing connections in the pool for the same target are destroyed too.

This is false by default, and only true for the CassandraConnectionPool. This
means that when one bad Cassandra connection is found, the others are
proactively destroyed.