Cassandra HA gets - only do consistency level ONE the first time
An operation may be performed against two different Cassandra processes (if the
first one times out).

The first time, we want to try a consistency level TWO get request.

If the connection to the Cassandra process times out, we'll retry to another
Cassandra process. In that case, we want to go straight to a consistency level
ONE request, because we've already spent a chunk of our latency budget failing
to contact the first Cassandra, and we want to ensure that we have time to make
the ONE request.