Unset the LD_LIBRARY_PATH variable when loading the IO trap.
This is needed because the library relies other symbols in cpp-common that are provided by the program being debugged.
If the library didn't unset the LD_LIBRARY_PATH it would get inherited by any subprocess spawned by the program (e.g.
when launching gdb to gather a stack trace). But since the subprocess doesn't have the relevant symbols it won't
launch correctly.